fix: PR #418 R2 — context hook 분리 + flicker 가드 + drift 트랩 제거
R2 review (Blocker 2 + medium 1 + question 1):
- SymbolChatContext.tsx의 useSymbolChat / usePublishSymbolChat을 hooks/useSymbolChat.ts로 분리 (CONVENTIONS.md hooks/ subfolder 규칙)
- usePublishSymbolChat의 useEffect를 publish/clear 두 개로 분리 — 매 state 변경마다 clear가 실행되며 발생하던 flicker 제거
- DependencyProgress의 "3개 축" 하드코딩을 {total} 참조로 교체 — AXIS_ORDER drift 트랩 (MISTAKES.md #15)
- useNewsAugment에서 enabled:false 대신 queryFn:skipToken 사용 — TanStack Query v5 idiomatic
- Suggestion 1건은 follow-up 이슈로 트래킹, Question 1건은 useSuspenseQuery 기반이라 안전함을 코멘트 회신
